Remove JournalDB usage from ChainDB
Fixes 607
